pyproject.toml
# Python project metadata, including dependencies & tool settings
# https://snarky.ca/what-the-heck-is-pyproject-toml

[project]
name = "rcv"
version = "0.1"

dependencies = [
    "attrs",
    "click",
    "prettyprinter",
    "zrp@git+https://github.com/egnor/zrp",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "pandas-stubs",
    "pip>=23.3.1",  # get past "error checking the latest version of pip" bug
    "visidata",
]

[project.scripts]
ea_zrp = "ea_race.ea_zrp:main"
zrp_test = "ea_race.zrp_test:main"
nb_to_ea_people = "nb_to_ea.people:main"
nb_to_ea_financial = "nb_to_ea.financial:main"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[tool.mypy]
namespace_packages = true
explicit_package_bases = true  # https://github.com/python/mypy/issues/13392

[tool.pylint.messages_control]
enable = ["useless-suppression"]
disable = [
    "fixme",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-public-methods",
    "too-many-statements",
]

include-naming-hint = "y"
variable-rgx = "([^\\W\\dA-Z][^\\WA-Z]*)$"
argument-rgx = "([^\\W\\dA-Z][^\\WA-Z]*)$"

[tool.setuptools.package-dir]
ea_race = "ea_race"
nb_to_ea = "nb_to_ea"
rcv_results = "rcv_results"
